typst_table

Print a table in raw typst for pdf document rendering

Usage

typst_table(data, linebreak = "<br>")

Arguments

data

A data.frame to render

linebreak

Line break pattern to be subbed by typst line break

See also

Examples

data_with_linebreak <- data.frame(Example = c("$10^2$<br>$-->$", "#sym.checkmark"))
typst_table(data_with_linebreak)
#>  [1] "```{=typst}"                "#table("                   
#>  [3] "  stroke: 1pt,"             "  align: horizon,"         
#>  [5] "  columns: 1,"              "  table.header[*Example*],"
#>  [7] "  [$10^2$ \\ $-->$], "      "  [#sym.checkmark], "      
#>  [9] ")"                          "```"